+++ src/constants/editor.ts
@@ -4,11 +4,11 @@ import { CodeEditorSuggestionItem, CodeEditorSuggestionItemKind } from '@grafana
  * Supported Languages
  */
 export const enum CodeLanguage {
-  SCSS = 'scss',
   HANDLEBARS = 'handlebars',
   HTML = 'html',
   JAVASCRIPT = 'javascript',
   MARKDOWN = 'markdown',
+  SCSS = 'scss',
 }
 
 /**
@@ -86,7 +86,7 @@ export const HelpersEditorSuggestions: CodeEditorSuggestionItem[] = [
  * Editor Type
  */
 export enum EditorType {
-  TEXT = 'text',
   HELPERS = 'helpers',
   STYLES = 'styles',
+  TEXT = 'text',
 }