+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / Enhancements
 
 - Update ESLint configuration (#192)
+- Increase Tests Coverage (#194)
 
 ## 4.0.0 (2023-06-26)